Allow grids to be specified by spacing or shape
Include option to give a spacing instead of a shape.
If this is the case, the spacing or the region might need to be adjusted
when the region is not a multiple of the spacing. If the region is
fixed, then the spacing is adjusted to the nearest multiple (default).
If the spacing is fixed, the East and North boundaries are adjusted.

Fixes #19