Functions to download sample data
Adds functionality to download data from Github or copy from a local
cache (using the env variable `VERDE_DATA_STORE`). Data is stored in
`$HOME/.verde/data`. When getting the data file name, check the hash
against the one stored on Github. If the hash is different, download
the file again to update it.