Add function to reduce data in blocks
Function `block_reduce` applies a reduction operation to data in blocks.
Also adds functions `get_region` to get the region from coordinates and
`inside` to get check which points are inside a given region. `inside`
allows passing in output and temporary arrays to avoid the overhead of
allocation in `block_reduce`.

Fixes #24