Add optional 'projection' arg to gridder methods (#31)
The projection argument should be a function that converts lat, lon
into Cartesian easting, northing `projection(lon, lat) -> easting, northing`
or between any two coordinate systems. This argument can be passed
to `grid`, `scatter`, and `profile`. The projection is applied to the 
generated coordinates before passing them to `predict`.
This way, we can fit a gridder on Cartesian data and specify a grid
in lat, lon. `fit` still takes in Cartesian data so the user needs to 
project the coordinates beforehand. 

Add an example of doing all of this using `ScipyGridder`.

Fixes #23