Add sample magnetic data from Rio de Janeiro
Data are a section of a survey from the 70s. Downloaded and cropped from
the [Geologic Survey of Brazil (CPRM)](http://geosgb.cprm.gov.br).
It's a good testing data set because it's dense and there is a regional
trend to the data. The signals from a few post-rift granitoids are
clearly visible in the data.