Use sphinx-gallery for the tutorials
Setup the tutorials as sphinx-gallery notebook style examples. This way
they can be run as normal Python files or downloaded as notebooks. They
play nicely with git, unlike ipynb files, and are run whenever the docs
are built. They can also easily link with the API docs and use rst
syntax like the rest of the docs.