Use 'coordinates' instead of individual arrays
All fit and predict methods have been changed to use a `coordinates`
tuple instead of `easting, northing`. This will allow the classes to be
used in more generic methods that take in more than 2 coordinates, like
equivalent layers.

Move the `base` subpackage back to a module and import `BaseGridder` in
the main API.

Needed for #42