Add Chain estimator (#44)
The `Chain` estimator takes a list of gridders. Each gridder will be fit
on the residuals of the previous one. For example, can be used to chain
trend estimation with gridding.

For a gridder to be compatible, it has to set a `residual_` attribute
in `fit` that contains `data - self.predict(coordinates)`. This is used
as input to the next gridder.

The interface is similar to the scikit-learn `Pipeline` estimator but
does not yet support everything (like accessing attributes from the
components).

Fixes #42