BaseGridder.filter method for use in Chain (#57)
Instead of relying on the `residual_` argument, implement a new `filter`
method in the `BaseGridder` that fits a model and returns the
coordinates, residuals, and weights. This can be used by `Chain` to
chain the estimators. The advantage of not relying on `fit` and
`predict` is that other operations like `block_reduce` can be used with
it (didn't fit the old model because it changes the coordinates as well).