Add BlockMean with options to output weights (#63)
Created a dedicated `BlockMean` class that can calculate blocked
weights using the data variance or propagating uncertainties.
Moved the block reduce code to a dedicated module because it's 
getting big. 
Add examples of the different types of weights and use it with a 
weighted spline on the GPS sample data.
Add a `variance_to_weights` function that can handle zeros and
NaNs in the variance and applies a cutoff to small variances.