Functions for cross-validation (#69)
Implements the `cross_val_score` and `train_test_split` functions 
with a similar interface as scikit-learn but that can act on our 
multi-component spatial data with optional weights.
`cross_val_score` can take a `dask.distributed.Client` and submit
tasks to it instead of doing the actual calculations. In this case,
the scores returned are futures and can be used in other distributed
calculations. If no client is given, then the computations are done
serially. Created a `DummyClient` that just executes the function
immediately to simplify the code.

Adds the `region` option for `verde.Spline` so that it can be used with
`cross_val_score`. Otherwise, model region would change for 
different data folds.

Fixes #55