Use pooch to manage the sample data (#98)
The Pooch package manages all that we had in `datasets/download.py` module
but in a much more sophisticated way. Use it instead.
Renames some of the data files, removes the stored hash files and keep them
in `verde/datasets/registry.txt` instead.
Document the usage and how to change the data dir in `example-data/README.txt`.
Replace the `pooch.__version__` variable with a `verde.version` module that defines
`full_version` and `git_revision` like numpy does. The dunder variable was a bit 
confusing and was causing problems to import on the datasets module.