Add recipe for simple gridding with Spline
There are no weights on this one so use the air temperature data from
Texas.
Renamed some of the recipes to make the order better.

Running the docs linkcheck only on the Travis Mac builds because they
sometimes fail through no fault of ours. This way, we can force merge
PRs and still have the docs deploy on master.