Use numba to optimize Sline.jacobian (#110)
Calculating the Jacobian using numba with parallel `for` loops results 
in 10x speedup compared to numpy. This translates to a 40-45% 
speedup in the whole process.
Keep numba as optional and only use it when installed or requested.
Refactor some of the underlying spline code to minimize repetition.
Move least-squares fitting to a separate function to be shared by all
classes.
Move methods shared between `Spline` and `VectorSpline2D` to 
functions and remove inheritance.