Simplify the spline arguments for force locations (#118)
Remove the arguments for laying out forces on a regular grid. This
reduces the complexity. Instead, force coordinates can be an argument to
the class. Also, force coordinates are fixed once set so the model won't
change configuration when fitting on different data. This was a major
problem for Kfold CV and difficult to detect if you don't know what's
happening. This way, it's explicit. One can still reset the forces by
re-instantiating the interpolator or setting `force_coords` to `None`.

Affects Spline and VectorSpline2D.