Use the README as the index page for the docs
The text in the README and the index page was duplicated. The only
difference was the header part.
Include the README source in `doc/index.rst` instead.
To have different headers, adds a comment to the README marking where to
start importing the source.
Also adds a list of links to previous documentation builds to the end of
the README.