Add fields to Rio Mag data to separate tie lines
Add the `line_type` and `line_number` fields to the Rio magnetic data so
that we can separate the regular flight lines from the tie lines (the
ones the cut across the survey E-W). This will be needed if we want to
blockmean the data without averaging cross-over points (they were
measured at different heights).