Simplify distance_mask and add a projection arg (#125)
`distance_mask` now can take grid coordinates or an `xarray.Dataset` 
to create the mask. If the `Dataset` is given, the coordinates are taken 
from it and the mask is applied to it. Otherwise, return the mask. If a 
projection is given, use it on the coordinates before calculating distances.