Flake8 check, simplify Makefile and Travis builds
Add flake8 to the `make check` command to get a quick quality check
(pylint is slow). Separate pylint on `make lint` which can be run less
often.

Took the opportunity to rename a few make commands:
* `develop` is now `install`, which is what it actually does
* `test` and `coverage` are merged because it doesn't hurt to always
  calculate coverage. This avoids repeated code in the Makefile.

Replace the Python 3.5 OSX Travis build with a Linux build that only
runs the style checks. This will be fast and quickly alert to style
problems while the other builds continue.