Add extra_coords option to coordinate generators (#145)
The `grid_coordinates`, `scatter_points`, and `profile_coordinates` all
accept a new a `extra_coords` argument with the constant value to be
used as an extra coordinate. If given, then the output will have an
extra coordinate array of the same shape as the rest and with that
single value. If a list is given, multiple extra coordinate arrays will
be returned.

This can be used to specify constant coordinates to a grid generator,
like height or time, to generate a 2D grid/scatter/profile from data
that requires 3+ coordinates.

A step towards #138