Don't use the Jacobian for prediction (#149)
All gridders create a Jacobian matrix to make predictions (using a dot product with the estimated
parameters). This is very inefficient with regards to memory. Generating largeish grids requires a
huge amount of RAM.

Replace these with a for loop summation. Accelerate them with numba to achieve the same 
performance. Refactor the Green's functions calculations into separate functions that can be
JITed. The pure Python version is slower than the dot product but at least it won't blow up 
memory.