Use cross_val_score in model selection tutorial
The previous version used the `train_test_split`, which was highly
sensitive to the random number seed. Update the tutorial to show this
effect and use `cross_val_score` instead for a better metric. Use only a
single `Spline` instead of the `Chain` because it's not needed for this
dataset.