Utility function to create kdtrees
We support using pykdtree as a replacement for most kdtree operations
but the feature set is the not identical to the scipy implementation.
We had no way of dealing with this before and had some repeated
`try...except` code for importing the modules.

A new `kdtree` function in `verde/utils.py` centralizes the imports,
allows using the scipy kdtree even if pykdtree is installed,  and
handles transposing the coordinate arrays before creating the tree.