New median_distance func for near neighbors  (#163)
The new `median_distance` function calculates the median near neighbor distance between
each point in the given dataset. It automates the process of creating a kdtree and
querying it on the same input points. It will take a median of the *k* nearest neighbors
for each point and handle projections and reshaping of the output.

This type of function is useful to get a sense for how scattered are the data points and
what would be a reasonable grid spacing or block mean spacing.

Fixes #127