Assign authorship to "The Verde Developers"
Replaces my name with "The Verde Developers" in authorship claims and
copyright notices. Include a definition of what that term means in the
`AUTHORS.md` file and include a reference to it in the README and doc
page.
Left my name and email as maintainer on the `setup.py` because we don't
have a project email yet.