header parse test
this creates a fake header as a list then passes it to a helper function that is the first half of the load_surfer function. It returns the header information after parsing it and compares to the original list of strings. Definitely a WIP as I figure out how to write tests. Any feedback is more than welcome.