Move to verde.io and refactor
Moved the function to a new verde.io module to separate it from the misc
utilities.
Refactor to use shape, region, etc instead of individual values.
Add support for paths or open files (or file-like objects).
Use test fixtures and test with and without nans.
Test the full function as well as the individual parts.
Fix formatting and linting errors.