Fix wrong coordinates for pixel-register and shape (#183)
Fix a bug in `grid_coordinates` when passing `pixel_register=True` and
`shape` instead of `spacing`. The returned coordinates had 1 too few
elements in each dimension (and the wrong values). This is because we
generate grid-line registered points first and then shift them to the
center of the pixels and drop the last point. This only works when
specifying `spacing` because it will generate the right amount of
points. When `shape` is given, we need to first convert it to
"grid-line" shape (with 1 extra point per dimension) before generating
coordinates.