Add SplineCV: a cross-validated Spline (#185)
The `SplineCV` class performs grid search cross-validation to auto-tune a `Spline`.
It uses `cross_val_score` to cross-validate and can be configured to run in parallel
with a `dask.distributed.Client` and take different scikit-learn splitters besides
`KFold`. Tested on the checkerboard function and add a gallery example using the
Texas temperature data.