Bug fix: Reset force coords when refitting splines (#191)
Previously, the `Spline` set the force coordinates from the data
coordinates only the first time `.fit` was called. This means that when
fitting on different data, the spline would still use the old force
coordinates. Now, the spline will use the coordinates of the current
data passed to `fit`. This only affects cases where `force_coords=None`.
It's a slight change and only affects some of the scores for
cross-validation.