Fix breaking Spline tests: extra warnings and type tests (#194)
Some tests rely on checking if the code gives a warning. The test was
checking if the number of warnings is 1 and comparing the error message.
This breaks if any of our dependencies introduce warnings.
Fix by checking if >=1 warning and if any have the right message.
Other doctests were printing the `type` of a variable which renders differently
sometimes. Best to not check these at all.