Add deprecation warning for Rio sample data
Licensing for the dataset is unclear it's best to remove it. We'll
deprecate first and then remove in a future major release. Add warnings
to the docstrings and sample data example. Replace the Rio data with
Texas wind in the Trend gallery example (it's already used in the trend
estimation tutorial). See #201 for more context.