Remove attributes argument from check_is_fitted
The attributes argument for the sklearn.utils.validation.check_is_fitted
function is going to be deprecated. Remove the attributes passed to the
check_is_fitted function used on the predict method of every predictor
from Verde. Modify the example on verde.base.BaseGridder to show how the
check_is_fitted function must be used when subclassing the BaseGridder.