Use pooch.os_cache and pkg_resources in datasets
With Pooch 0.7.0, the recommended way of loading the registry file is
with `pkg_resources` (see fatiando/pooch#120). It's also better to use
the default cache location so users can more easily clean up unused
files. Because this is system specific, add the `verde.datasets.locate`
function to return the cache folder location.