Use dask.delayed instead of the distributed client (#222)
It's easier and allows building the entire graph lazily before
executing. In cross_val_score, the lazy results are returned and can be
joined with other lazy operations before computing. SplineCV computes
scores lazily to triggers computation to get the best result. Add Dask
as a dependency for this to work. Deprecate the 'client' argument (to be
removed in 2.0.0). Use delayed in the tutorials.