Explicitly set numba config for each jit function (#221)
Using `parallel=True` on the Green's function was raising a warning
because there is nothing to parallelize. Set arguments individually to
avoid this. Add missing `numba.prange` to Spline Jacobian calculation to
actually run it in parallel. A combination of Dask + numba makes some 
tests hang on CI randomly even if passing locally.Run the tests that use 
dask with the numpy engine to avoid this.