Fix bug with profile distances being unprojected
`BaseGridder.profile` has the option to apply a projection to the
coordinates before predicting so we can pass geographic coordinates to
cartesian gridders. In these cases, the distance along the profile is
calculated by `profile_coordinates` with the unprojected coordinates (in
the geographic case it would be degrees). Returning unprojected
distances is not very useful and instead we should have been returning
projected distances. They need to be calculated from the projected
coordinates (out of `profile_coordinates`) because the points are evenly
spaced in unprojected coordinates.