Add function to project gridded data
The project_grid function transforms a grid to the given projection.
Resamples using ScipyGridder (by default) and runs a blocked mean
(optional) to avoid aliasing when the points aren't evenly distributed
in the projected coordinates (like in polar projections).
Move project_region and this new function to verde/projections.py.
Applies a convexhul_mask to the grid always to avoid extrapolation to
points that had no original data. Only works for DataArray because there
is no easy way to guarantee that a Dataset variables have the same dims.