Add a blocked K-Fold cross-validator
The `BlockKFold` class splits the data into spatial blocks and does
K-fold cross-validation splits on the blocks. Unlike `BlockShuffleSplit`
it's difficult to balance the folds to have the same amount of data
(since blocks can have different numbers of points) so each fold will
have the same number of blocks but not necessarily data. Refactor code
from `BlockShuffleSplit` into a base class that can be shared with the
new class.