Add function to find partition points of equal sum
The function finds the indices where an array should be split to have n
parts of approximate equal sum. Sometimes this can't be done (there
would be empty partitions) so raise an exception when that happens. It's
more likely to happen when the number of parts is close to the number of
elements.