Add option to pass coordinates to the grid method
Start drafting the addition of an optional coordinates argument to the
grid method of the BaseGridder.
It would allow use to predict on preexisting grids and obtain
a xr.Dataset as output.
Add a new example that passes the coordinates of a preexisting grid to
a Chain with a Spline inside.
Write a new meshgrid_from_1d function that creates a meshgrid out of 1d
horizontal arrays, but also checks if they have a single dimensions and
supports for passing extra coordinates as 2d-arrays.