Add option to pass coordinates to the grid method (#326)
Allow `BaseGridder.grid` to take the coordinates of the preexisting grid either
as 2d arrays (meshgrid) or as 1d arrays. Raise a `FutureWarning` if the
`spacing`, `shape` or `region` arguments are passed, since they will be removed
in v2.0.0. Write a new `meshgrid_from_1d` function that creates a meshgrid out
of 1d horizontal arrays, checks if they have a single dimension, and supports
for passing extra coordinates as 2d-arrays.