Move the sources of gallery/tutorials to doc (#354)
It was strange that the sources for those was not in the doc folder,
even though they are only meant to be viewed in the documentation. Move
them to the doc folder to reduce the number of things in the top level
of the repository. Use the suffix `_src` to distinguish the source files
for these galleries from the generated galleries, retaining the current
URLs.