Modernize the front page of the docs (#356)
Use sphinx-panels to make the 4 box callouts to sections of the docs.
Stop reusing the README and instead have a single paragraph explaining
what Verde is. Reuse this paragraph in the README since that one needed
some updating as well. Move the links to other versions of the docs into
it's own page. Rename the API reference to something more informative.
Remove the deprecated `test` function from the API reference.