New logo and use sphinx-design in the docs (#367)
Created a new logo for Verde: A mix between an avocado and a
contour map. The color scheme is based on viridis, which was
the inspiration for the original logo. This works better as a
an actual logo as opposed to the text version. Took this
opportunity to add it to the front page using sphinx-design as
a replacement for sphinx-pannels. Not a major change.