Set lower bounds for dependencies based on NEP29 (#384)
Support all minor versions of Python from the past 42 months and all
minor versions of dependencies from the past 24 months. Add a document
to the docs explaining the policy. Add jobs to the CI to test oldest and
newest versions and run on oldest Python and dependencies and newest 
Python and dependencies instead of every combination of both. This 
covers the functionality that we expect to support and produces many 
fewer workflow runs. Update the cache action to v3 as well.