Use PyGMT instead of Cartopy in Chain tutorial
The PyGMT maps require less code and tend to look nicer than the Cartopy
ones. It's also faster for non-rectangular maps. Add PyGMT and GMT as
docs dependencies with pinned versions to avoid breaking the docs build.
Replacing in a single tutorial to establish the infrastructure. Replace
in other places in follow-ups.

Co-authored-by: Santiago Soler <santisoler@fastmail.com>