Add line_coordinates, a 1D version of grid_coordinates (#390)
Instead of adding the ability to generated 3D coordinates to
grid_coordinates, it's more straightforward to add a 1D version of
numpy.linspace that can take a spacing instead of a number, knows how to
adjust spacing/region, and can return pixel-registered coordinates. This
function is then used in grid_coordinates and can be used to generate
n-dimensional coordinates if desired (by calling numpy.meshgrid).