Fix behavior if spacing >= 2 * interval (#406)
This was a bug caused by the rounding when calculating the number of
nodes in `spacing_to_size`. For the case where the spacing is larger
than twice the interval, the size calculating rounded to zero which
meant that we were getting a single value for the number of nodes. This
is clearly wrong since our behavior is to change the spacing or the
region to make sure we always produce at least 2 nodes in the interval.