Remove dependence on pkg_resources (#448)
We used it to get the registry file from the package data. But the file
only had 4 entries so it made no sense to do all of this instead of
having the registry as a dict directly in the code. So remove the
registry file from the package and then we don't need pkg_resources.