Replace Sphinx napoleon for numpydoc (#450)
Replace the built-in Sphinx plugin `napoleon` for `numpydoc`. This
solves the issue of attributes with trailing underscore not rendering
properly in docs. Add `numpydoc` as requirement for building the docs
and add it to the Sphinx configuration file. Remove `napoleon` from
Sphinx configuration. Remove the "Methods Summary" section in the Sphinx
class template since `numpydoc` builds its own list of all methods.