Exploit the symmetry of the spline Jacobian
When nodes aren't given and are the same as the data coordinates, the
Jacobian is symmetric and we can calculate only the upper or lower
triangular part. In serial computation, exploiting this leads to ~50%
decrease of computation time. In parallel, numba does some kind of
optimization that doesn't work on the symmetric implementation and the
symmetric code is actually a little bit slower than the naive
implementation. So only use the new code when running the Jacobian
building in serial. This doesn't have a huge impact on performance but
it's significant enough.