Remove block_split and rename spacing arguments (#498)
Remove the `block_split` function which moved to Bordado. Renamed the
`spacing` arguments in `BlockReduce`, `BlockMean`, `BlockKFold`,
`BlockShuffleSplit` and `train_test_split` to `block_size` and `shape`
to `block_shape`. This is the new argument name in `bordado.block_split`
and it makes much more sense. Update tests and examples to match. Add
Bordado as a dependency.