Fix sphinx warnings about autosummary stub files
The combination of autosummary and numpydoc was causing warnings about
missing stub files because numpydoc was building a toctree for methods
but the pages didn't exist (they are all in the main page). Disable this
option and remove the attribute autogenerated pages since we already
describe important attributes in the docstring.