Remove the Chain composite estimator (#511)
The Chain class was introduced to help perform trend estimation and
spline interpolation in a single step. This was to help prevent data
leakage from the trend estimation step. In practice, this is never
really used since the trend ended up not being that necessary. The
existence of Chain makes it necessary to have the `filter` method, which
is awkward. It also makes it necessary for BlockMean and BlockReduce to
be classes with a single method (filter) instead of functions, as would
be more natural. Removing the Chain fixes all of these issues and won't
have a large impact on usability. With the Chain removed, the filter
method of the BaseGridder can also be removed.