Add `percentile` parameter to `maxabs()` function (#524)
Allow returning a percentile of the absolute value of the provided
values, instead of just the maximum absolute value. For example, using
the 95% percentile instead of the maximum will create a colorbar that is
less washed out and reduces the effects of outliers in the data.
If multiple arrays are provided, they need to be concatenated before
calculating the quantiles, so it can be significantly slower than just
the maximum absolute value determination.

---------

Co-authored-by: Santiago Soler <santisoler@fastmail.com>